Login at /login and dashboard at /
